requirements.txt
click~=8.1.6
piexif~=1.1.3
pytest~=7.4.0
loguru~=0.7.0
Pillow~=10.0.0